+++ lib/getch/gentoo/chroot.rb
@@ -11,7 +11,7 @@ module Getch
         return if STATES[:gentoo_update]
         puts "Downloading the last ebuilds for Gentoo..."
         Helpers::create_dir("#{MOUNTPOINT}/var/db/repos/gentoo")
-        cmd = "emerge-webrsync"
+        cmd = "emaint sync --auto"
         exec_chroot(cmd)
       end